refactor(inventory): isolate exported record bindings
(cherry picked from commit d59927c0218fdf5725faa94f0f1e54574f3385fd)
