+++ scripts/nemo/lib/inventory-bindings.cjs
@@ -6,6 +6,7 @@ const {
   parameterIntact, paramLookupRe, paramLookups, returnsOf, subjectStartAt,
   matchClose, functionBodyAt,
 } = require('./inventory-lexer.cjs');
+const { recordExportedRecordArrayBindings } = require('./inventory-records.cjs');
 
 function createBindings(idx, domRows, rel) {
   const { allFns, fnIndex, resolveFn } = createFunctionIndex(idx);
@@ -342,82 +343,11 @@ function createBindings(idx, domRows, rel) {
       }
     }
   }
-  // A window-exported helper can receive an array of literal records.  Follow
-  // only an exact `window.Object.member(...)` call whose member is proven by
-  // the matching object-literal export, then use the helper's own lookup of
-  // `record.property`.  This keeps a same-named helper or a record-shaped
-  // string from creating a binding.
-  function exportedMember(objectName, memberName) {
-    const re = new RegExp('\\bwindow\\.' + esc(objectName) + '\\s*=\\s*\\{', 'g');
-    for (const exportFile of idx) {
-      re.lastIndex = 0;
-      let m;
-      while ((m = re.exec(exportFile.code))) {
-        const open = exportFile.code.indexOf('{', m.index);
-        if (exportFile.kind[m.index] !== CODE || open < 0 || exportFile.kind[open] !== CODE) continue;
-        for (const prop of argsOf(exportFile, open).args) {
-          const value = new RegExp('^' + esc(memberName) + '\\s*:\\s*([A-Za-z_$][\\w$]*)\\s*$').exec(prop.text);
-          if (value) {
-            const fn = resolveFn(exportFile, value[1], prop.start);
-            if (fn) return fn;
-          }
-        }
-      }
-    }
-    return null;
-  }
-  function literalRecordProperty(f, objectOpen, property) {
-    for (const prop of argsOf(f, objectOpen).args) {
-      const value = new RegExp('^' + esc(property) + '\\s*:\\s*(.*)$').exec(prop.text);
-      if (value) return stringLiteral(value[1]);
-    }
-    return null;
-  }
-  function recordArrayBindings(fn, property) {
-    const out = EMPTY();
-    const list = fn.params[0];
-    if (!list) return out;
-    const re = new RegExp('(?:^|[^\\w$.])' + esc(list) + '\\s*\\.\\s*forEach\\s*\\(', 'g');
-    let m;
-    while ((m = re.exec(fn.f.code.slice(fn.start, fn.end + 1)))) {
-      const callAt = fn.start + m.index;
-      if (fn.f.kind[callAt] !== CODE) continue;
-      const open = fn.f.code.indexOf('(', callAt);
-      const cb = callbackOf(fn.f, argsOf(fn.f, open).args[0] || { text: '' });
-      if (!cb || !cb.param0 || !cb.span) continue;
-      const lookup = new RegExp('(?:window\\.)?document\\.getElementById\\(\\s*' + esc(cb.param0) + '\\s*\\.\\s*' + esc(property) + '\\s*\\)', 'g');
-      let hit;
-      while ((hit = lookup.exec(fn.f.code.slice(cb.span.start, cb.span.end + 1)))) {
-        const at = cb.span.start + hit.index;
-        if (fn.f.kind[at] !== CODE || !parameterIntact(fn.f, cb.span, cb.param0, at)) continue;
-        const openAt = fn.f.code.indexOf('(', at);
-        merge(out, resolveLookup(fn.f, subjectStartAt(fn.f, at), matchClose(fn.f, openAt) + 1, newCtx(false, 'exported record helper')));
-      }
-    }
-    return out;
-  }
-  const EXPORTED_RECORD_CALL_RE = /window\.([A-Za-z_$][\w$]*)\.([A-Za-z_$][\w$]*)\s*\(/g;
-  for (const f of idx) {
-    EXPORTED_RECORD_CALL_RE.lastIndex = 0;
-    let m;
-    while ((m = EXPORTED_RECORD_CALL_RE.exec(f.code))) {
-      if (f.kind[m.index] !== CODE) continue;
-      const fn = exportedMember(m[1], m[2]);
-      if (!fn) continue;
-      const open = f.code.indexOf('(', m.index);
-      const first = argsOf(f, open).args[0];
-      if (!first || f.code[first.start] !== '[' || f.kind[first.start] !== CODE) continue;
-      const arrayClose = matchClose(f, first.start);
-      if (arrayClose + 1 !== first.end) continue;
-      const res = recordArrayBindings(fn, 'wrap');
-      if (!res.bindings.length) continue;
-      for (const record of argsOf(f, first.start).args) {
-        if (f.code[record.start] !== '{' || f.kind[record.start] !== CODE || matchClose(f, record.start) + 1 !== record.end) continue;
-        const id = literalRecordProperty(f, record.start, 'wrap');
-        if (id !== null) addRec(bindings, id, recFrom(f, m.index, res, 'exported ' + m[1] + '.' + m[2] + '()'));
-      }
-    }
-  }
+  recordExportedRecordArrayBindings({
+    idx, CODE, esc, argsOf, stringLiteral, bindingOwner, resolveFn, callbackOf, parameterIntact,
+    subjectStartAt, matchClose, EMPTY, merge, resolveLookup, newCtx,
+    addRec, bindings, recFrom,
+  });
   // Lookups that go through a table of id literals — an object or an array,
   // flat or of objects — either directly (`getElementById(IDS[mode])`,
   // `getElementById(IDS.unite)`) or through the parameter of a callback that
@@ -440,15 +370,16 @@ function createBindings(idx, domRows, rel) {
       const ids = uniq((lit.match(/'([\w-]+)'|"([\w-]+)"/g) || []).map((s) => s.slice(1, -1)).filter((s) => domIds.has(s)));
       if (!ids.length) continue;
       const name = m[1];
-      addRoot(f, { name, ids, start: 0, end: f.code.length });
+      if (f.scopes.some((s) => s.function && s.params.includes(name) && s.start <= m.index && m.index <= s.end && bindingOwner(f, name, m.index) !== s)) continue;
+      addRoot(f, { name, ids, start: 0, end: f.code.length, owner: bindingOwner(f, name, m.index) });
       const iter = new RegExp('(?:^|[^\\w$.])' + esc(name) + '\\s*\\.\\s*(?:forEach|map|some|every|filter)\\s*\\(', 'g');
       let c;
       while ((c = iter.exec(f.code))) {
         const cb = callbackOf(f, argsOf(f, c.index + c[0].length - 1).args[0] || { text: '' });
-        if (cb.param0 && cb.span && cb.f === f) addRoot(f, { name: cb.param0, ids, start: cb.span.start, end: cb.span.end });
+        if (cb.param0 && cb.span && cb.f === f) addRoot(f, { name: cb.param0, ids, start: cb.span.start, end: cb.span.end, owner: bindingOwner(f, cb.param0, cb.span.start) });
       }
       const forOf = new RegExp('for\\s*\\(\\s*(?:var|let|const)\\s+([A-Za-z_$][\\w$]*)\\s+of\\s+' + esc(name) + '\\s*\\)', 'g');
-      while ((c = forOf.exec(f.code))) { const span = functionBodyAt(f, c.index + c[0].length); if (span) addRoot(f, { name: c[1], ids, start: span.start, end: span.end }); }
+      while ((c = forOf.exec(f.code))) { const span = functionBodyAt(f, c.index + c[0].length); if (span) addRoot(f, { name: c[1], ids, start: span.start, end: span.end, owner: bindingOwner(f, c[1], span.start) }); }
     }
   }
   const TABLE_LOOKUP_RE = /getElementById\(\s*([A-Za-z_$][\w$]*)\s*[[.]/g;
@@ -459,7 +390,7 @@ function createBindings(idx, domRows, rel) {
     TABLE_LOOKUP_RE.lastIndex = 0;
     while ((m = TABLE_LOOKUP_RE.exec(f.code))) {
       if (f.kind[m.index] !== CODE) continue;
-      const roots = (tableRoots.get(f.file) || []).filter((r) => r.name === m[1] && r.start <= m.index && m.index <= r.end);
+      const roots = (tableRoots.get(f.file) || []).filter((r) => r.name === m[1] && r.owner === bindingOwner(f, m[1], m.index) && r.start <= m.index && m.index <= r.end);
       if (!roots.length) continue;
       const rec = recordAt(f, m.index);
       for (const id of uniq(roots.flatMap((r) => r.ids))) addRec(bindings, id, rec);
